backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain-core>=0.3.49,<1.0.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-groq==0.3.2
slowapi==0.1.9